requirements.txt
flask-appconfig>=0.10
flask-bootstrap
flask-debug
flask-wtf
Flask-SQLAlchemy
